chore: moved generated global types to src
