+++ scripts/build-schema.js
@@ -3,6 +3,6 @@ import path from 'path'
 import { printSchema } from 'graphql/utilities'
 import { schema } from '../src/api/schema'
 
-const outputPath = path.resolve(__dirname, '../generated/schema.graphql')
+const outputPath = path.resolve(__dirname, '../src/generated/schema.graphql')
 
 fs.writeFileSync(outputPath, printSchema(schema), 'utf8')
