+++ src/routes/NodeProjectDetails/graphql-types/PROJECT_QUERY.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { SemverOutdateStatus } from "./../../../config/graphql-types";
+import { SemverOutdateStatus } from "./../../../generated/graphql-types";
 
 // ====================================================
 // GraphQL query operation: PROJECT_QUERY
